feat(Blockchain) - #624 - added shared package of blockchainAddresses to contracts folder. Key was to split blockchainAddresses into core addresses and then Viem import. Viem is the dependency that allows us to write things to change. But contracts doesn't write things to chain, doesn't need it where as frontend/functions do. So we split them up and then just copy the blockchainAddresses.core over to contracts while the others get teh full blockchainAddresses including the viem package
