+++ contracts/package.json
@@ -7,7 +7,8 @@
     "test": "hardhat test",
     "deploy:sepolia": "hardhat run scripts/deploy.ts --network sepolia",
     "deploy:base": "hardhat run scripts/deploy.ts --network base",
-    "verify": "hardhat verify"
+    "verify": "hardhat verify",
+    "copy-shared": "tsc ../packages/shared/src/blockchainAddresses.core.ts --outDir scripts/_shared --module es2022 --target es2022 --moduleResolution bundler --skipLibCheck --declaration false"
   },
   "dependencies": {
     "@account-abstraction/contracts": "^0.7.0",
